(build) OpenBSD warning fixes
FossilOrigin-Name: 10f0c9a7526a133b8c5f6f8d08da85311f4adcace305b333721cfae54f8239dc

+++ src/platform/posix/map_resource.c
@@ -26,7 +26,7 @@
 #endif
 
 static inline bool read_safe(
-	int fd, size_t ntr, size_t bs, char* dofs, ssize_t* outsz)
+	int fd, size_t ntr, size_t bs, char* dofs, off_t* outsz)
 {
 	char emptyb[bs];
 	char* dbuf = dofs ? dofs : emptyb;
